requirements.txt
tensorflow
fastapi
uvicorn
python-multipart
streamlit
pandas
matplotlib
numpy
pillow
locust
scikit-learn
requests